api/requirements.txt
mysql-connector-python
requests
datetime
